final version. Here is the summary doc:
initialize a few variables. isdataframe = FALSE and others.
we check if user passed a string or an object
if object (non-string) isdataframe is still FALSE so we print error in the sink and return from the function
if string was passed, we check if it contains package name (e.g. 'datasets::mtcars')
if package name is passed we extract package-name and dataset-name in separate variables
now we check if package is loaded. if not we print error in sink and return.
if package is loaded, we get the environment of the package
now check if dataset-name (extracted above) is in the package environment.
if dataset-name is not found in the package environment we print error and return
if dataset-name is found, we check if it is really a data.frame class and set a flag 'isdataframe'.
if it isdataframe=TRUE, we make a copy of the dataset in the global-environment, for loading dataset in the later steps.
Now if user passed a string that does not have the package-name, we check if dataset exists in the global-environment.
if it exists we check if it is a data.rame class and set a flag 'isdataframe'
now we check if isdataframe is FALSE we print error message and return
to this point isdataframe is TRUE we execute the dataset loading steps.