type.convert in processNewDataset does not change class of column if it is non-character column.
putEmptyCellsBack has been modified: When dialog is closes on a partially
filled dataset (like Dataset1, Dataset2 etc.) after dialog execution(or cancel) we
put back empty rows and columns so that user can fill it further if needed.
But this was converting all the filled columns back to 'character'.
So if user has made some column factor or numeric etc and ran a dialog, after
dialog execution all columns change to 'character' type which is frustrating
for the user if he/she wants to run same or other dialog on the same
dataset(Dataset1 etc) they have to repeat the process of making col factor
or numeric etc.
This issue has been fixed. I used tibble::add_row() to add back empty rows
and the function will take care of filled column class when adding empty
rows (with NAs). Empty columns are still done the old way (no change).
Each column that is appended, is of type 'character'