fixes for empty-grid (eg Dataset1) for issues reported by Ross.
RULES:
1) If we rename variables, make variables factor and add factor levels but entered no data in the empty grid and now we save this grid, this grid will retain all the vairables with class/type etc. This will be handy when one user creates all the variables with required type and other user is supposed to fill that dataset.

2) Defining new dataset boundaries (with example):
Say you only enterted data in a cell that has row index = 5 and column index = 4. This cell will be considered as a last cell of the dataset. So your dataset will awaly start from the top-left first cell (row index = 0 and column index = 0) and ends on the row index = 5 and column index = 4. Only this part(i.e. [0,0] to [5,4]) of the data grid will be used while running any analysis or while saving the data grid. All the rows and columns beyond the last cell(i.e. [5,4] in this case) will be PERMANENTLY removed during SAVE and temporarily removed during analysis. Once the analysis finishes, all the removed rows and columns are added back to the data grid. All the rows and columns that fall under the dataset boudary are retained no matter if they are filled or not.
TIP : Before saving the new dataset, always mark the dataset boudary by picking up the desired last cell and filling a value in it.