if user convert 'Char to Date' and generates new col not the total col count should be 16 (before anaylsis it was 15). There was and issue where not all the dropped empty cols were not getting appended. If analysis generated one new col we were adding one less col (empty col) This has been fixed now.
Other changes:
now we process the empty grid as follows:
We find the index of the col that was renamed (check from right to left)
We also find the highest index of the col that has the data.
Now we have 2 indexes, whichever is greater we take that and drop rest of the empty cols thus creating subDataset (that can be saved or on which we can run analysis)
If user only renamed cols and enter no data then we pick the highest index col the was renamed and drop all other cols to the right of this renamed col. And this is our subDataset that can be saved but analysis can't be run because of no data.